# swagger.yaml
openapi: 3.0.3
info:
  title: GitHub Release Notifier API
  description: Subscribe to email notifications about new GitHub repository releases
  version: 1.0.0

servers:
  - url: http://localhost:3001
    description: Local development

paths:
  /health:
    get:
      summary: Health check
      tags: [System]
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /subscriptions:
    post:
      summary: Create subscription
      description: Creates a pending subscription and sends a confirmation email. No API key required.
      tags: [Subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, repo]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                repo:
                  type: string
                  example: anthropics/anthropic-sdk-node
      responses:
        '201':
          description: Subscription created (pending confirmation)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Validation error
        '409':
          description: Subscription already exists or confirmation email already sent

  /subscriptions/{id}:
    get:
      summary: Get subscription by ID
      tags: [Subscriptions]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Subscription found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '404':
          description: Subscription not found

    delete:
      summary: Delete subscription
      tags: [Subscriptions]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Subscription deleted
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '404':
          description: Subscription not found

  /confirm/{token}:
    get:
      summary: Confirm subscription
      description: Activates a pending subscription. Link is sent via confirmation email. Returns HTML page.
      tags: [Confirmation]
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Subscription confirmed (HTML page)
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Invalid or expired token (HTML page)
          content:
            text/html:
              schema:
                type: string

  /unsubscribe/{token}:
    get:
      summary: Unsubscribe
      description: Deactivates a subscription. Link is included in release notification emails. Returns HTML page.
      tags: [Confirmation]
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Unsubscribed successfully (HTML page)
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Invalid token (HTML page)
          content:
            text/html:
              schema:
                type: string

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        owner:
          type: string
        repo:
          type: string
        status:
          type: string
          enum: [pending, active, inactive]
        confirmationToken:
          type: string
          format: uuid
        tokenExpiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time